// src/types/about.ts
export interface AboutPage {
  id: string;
  titleEn: string;
  titleAr: string;
  taglineEn: string;
  taglineAr: string;
  bodyEn: string;
  bodyAr: string;
  missionEn: string;
  missionAr: string;
  visionEn: string;
  visionAr: string;
  email: string;
  phone: string;
  addressEn: string;
  addressAr: string;
  website: string;
  facebook: string;
  instagram: string;
  twitter: string;
  linkedin: string;
  tiktok: string;
  snapchat: string;
  whatsapp: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}